Added qodana and fixed issues

+++ src/main/java/visualizer/level/element/dynamicelements/MonsterVisualizer.java
@@ -2,7 +2,6 @@ package visualizer.level.element.dynamicelements;
 
 import model.elements.dynamicelements.Monster;
 import gui.GUI;
-import model.elements.dynamicelements.Player;
 
 import javax.imageio.ImageIO;
 import java.awt.image.BufferedImage;
@@ -35,6 +34,6 @@ public class MonsterVisualizer implements DynamicElementVisualizer<Monster> {
 
     private String getSprite(Monster monster) {
         List<String> spriteActionList = spriteMap.get(monster.getOrientation());
-        return spriteActionList.get((int) (monster.getLastControlCount()/5 % spriteActionList.size()));
+        return spriteActionList.get((monster.getLastControlCount()/5 % spriteActionList.size()));
     }
 }
